+++ db/upgrade.php
@@ -18,9 +18,9 @@
  * @see uninstall_plugin()
  *
  * @package       mod_dmelearn
- * @author        Kien Vu
- * @copyright     2015 Brightcookie.com {@link http://www.brightcookie.com.au}
- * @version       1.0.0
+ * @author        AJ Dunn, Kien Vu
+ * @copyright     2015 BrightCookie (http://www.brightcookie.com.au), Digital Media e-learning
+ * @version       1.0.1
  * @license       http://www.gnu.org/copyleft/gpl.html GNU GPL v3 or later
  */
 
@@ -29,13 +29,13 @@ require_once($CFG->dirroot . '/mod/dmelearn/lib.php');
 /**
  * Execute mod_dmelearn upgrade from the given old version.
  *
- * @param int $oldversion
- * @return bool
+ * @param int $oldversion the version being upgraded from
+ * @return bool was upgrade successful
  */
 function xmldb_dmelearn_upgrade($oldversion = 0) {
-    global $DB;
+    global $DB, $CFG;
 
-    // Check if version between 1.0.2 and 1.1.1.
+    // Check if old version was between v1.0.2 and v1.2.0.
     if ($oldversion >= 2015051200 && $oldversion < 2015072300) {
         // Fix missing dmelearn grades in gradebook by forcing full update of grades for
         // courses containing dmelearn activities.
@@ -52,6 +52,26 @@ function xmldb_dmelearn_upgrade($oldversion = 0) {
             foreach ($courseids as $courseid)
             {
                 grade_grab_course_grades($courseid->id, 'dmelearn');
+            }
+        }
+    }
+
+    // Clear the twig cache, if possible.
+    // Check if old version included twig cache directory (v1.1.0 or newer).
+    if ($oldversion >= 2015062900) {
+        // Get the Twig Cache folder directory.
+        $twigcache = $CFG->dirroot . '/mod/dmelearn/content/template_cache';
+        // Simple check to prevent wrong directory being set as twig cache.
+        $containscache = strpos($twigcache, '/template_cache');
+
+        // If twig cache directory is writeable delete the files inside it.
+        if (is_writable($twigcache) && ($containscache !== false)) {
+            // Clear each Twig Cache File.
+            foreach (new RecursiveIteratorIterator(new RecursiveDirectoryIterator($twigcache),
+                         RecursiveIteratorIterator::LEAVES_ONLY) as $cachefile) {
+                if ($cachefile->isFile()) {
+                    @unlink($cachefile->getPathname());
+                }
             }
         }
     }