+++ CHANGES.md
@@ -1,5 +1,12 @@
 # Changelog
 
+## v1.3.0 (2015102200)
+- Plugin now requires PHP 5.4.0 or above.
+- Upgrade Guzzle to 5.3 and Twig to v1.22.3.
+- Update course templates v1 and v2 to support newer DM e-Learning courses.
+- On upgrade the plugin now removes old cache files in Twig Cache directory.
+- Add Moodle PHPUnit tests for this plugin.
+
 ## v1.2.5 (2015080501)
 - Fix issue with null if none in dmelearn_update_grades() function being set as true by default.
 - $PAGE->activityrecord skipped in view.php as we don't have a record.